requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
dnspython
validate_email_address
validate_email
openpyxl
pandas
bs4
selenium
unidecode
urllib3
pycountry